Pack cone data into 4 bytes by using int8 instead of half

+++ ImperialEngine/ImperialEngine/Shaders/glsl/DescriptorSet1.h
@@ -56,7 +56,7 @@ struct ms_MeshData
 struct Meshlet
 {
     BoundingVolume boundingVolume;
-    float16_t cone[4];
+    int8_t cone[4];
     uint vertices[64];
     uint8_t indices[126 * 3]; // up to 126 triangles
     uint8_t triangleCount;
