+++ ImperialEngine/ImperialEngine/src/backend/VariousTypeDefinitions.h
@@ -73,7 +73,7 @@ namespace imp
 	struct alignas(16) Meshlet
 	{
 		BoundingVolumeSphere BV;
-		uint16_t cone[4];
+		int8_t cone[4];
 		uint32_t vertices[64];
 		uint8_t indices[126 * 3]; // up to 126 triangles
 		uint8_t triangleCount;
